dkls23::keygen
Function message_receivers

pub fn message_receivers<S, F>(setup: &S, msg_receiver: F)
where
    S: ProtocolParticipant,
    F: FnMut(MsgId, &S::MessageVerifier),

Generates a map of message receivers for the DKG protocol.

This function iterates through all other parties in the protocol and calls the provided closure for each message ID and corresponding verifier key. It handles all message types used in the DKG protocol, including abort messages and messages for each round.

Arguments

setup - The protocol participant setup containing party information

msg_receiver - A closure that will be called for each (message_id, verifier) pair

Message Types

The function handles the following message types:

ABORT_MESSAGE_TAG - Protocol abort messages

DKG_MSG_R1 - Round 1 messages

DKG_MSG_R2 - Round 2 messages

DKG_MSG_OT1 - Oblivious transfer messages

DKG_MSG_R3 - Round 3 messages

DKG_MSG_R4 - Round 4 messages