dkls23::keygen::keyshare
Struct Keyshare

pub struct Keyshare { /* private fields */ }

Key share of a party.

Implementations

impl Keyshare

pub const MAGIC: [u8; 4]

Identified of key share data

impl Keyshare

pub fn new(n: u8, t: u8, id: u8, extra: &[u8]) -> Keyshare

Allocate an instance of a key share with given parameters.

pub fn from_bytes(buffer: &[u8]) -> Option<Self>

Create a key share from slice of bytes.

pub fn from_vec(buffer: Vec<u8>) -> Result<Self, Vec<u8>>

Create a key share from a given vector. Returns a passed vector in case of an error.

pub fn as_slice(&self) -> &[u8]

Return underlying byte slice.

pub fn public_key(&self) -> ProjectivePoint

Return public key as ProjectiveProint.

pub fn rank_list(&self) -> Vec<u8>

Return vector of ranks

pub fn x_i_list(&self) -> Vec<NonZeroScalar>

x_i_list

pub fn zero_ranks(&self) -> bool

Return true if all parties has rank zero.

pub fn get_rank(&self, party_id: u8) -> u8

Return rank of a party.

pub fn s_i(&self) -> Scalar

Return the secret scalar s_i.

pub fn extra_data(&self) -> &[u8]

Return user defined data embedded into the key share. The data is passed via setup message at time of key generation and are immutable.

pub fn big_s(&self, party_id: u8) -> ProjectivePoint

Get the big S value of another party

impl Keyshare

pub fn root_chain_code(&self) -> [u8; 32]

Returns root chain code

pub fn root_public_key(&self) -> ProjectivePoint

Returns public key as ProjectivePoint

impl Keyshare

pub fn get_finger_print(&self) -> KeyFingerPrint

Get the fingerprint of the root public key

pub fn derive_with_offset( &self, chain_path: &DerivationPath, ) -> Result<(Scalar, ProjectivePoint), BIP32Error>

Get the additive offset of a key share for a given derivation path

pub fn derive_child_pubkey( &self, chain_path: &DerivationPath, ) -> Result<ProjectivePoint, BIP32Error>

Derive the child public key for a given derivation path

pub fn derive_xpub( &self, prefix: Prefix, chain_path: DerivationPath, ) -> Result<XPubKey, BIP32Error>

Derive the extended public key for a given derivation path and prefix

Arguments

prefix - Prefix for the extended public key (Prefix has commonly used prefixes)

chain_path - Derivation path

Returns

XPubKey - Extended public key