dkls23::keygen::quorum_change
Function message_receivers

pub fn message_receivers<S, F>(setup: &S, msg_receiver: F)
where
    S: QuorumChangeSetupMessage<Keyshare, ProjectivePoint>,
    F: FnMut(MsgId, &S::MessageVerifier),

Processes message receivers for the quorum change protocol.

This function handles the distribution of messages to the appropriate receivers based on the protocol setup and message type.

Type Parameters

S - A type implementing the QuorumChangeSetupMessage trait

F - A closure type for processing message receivers

Arguments

setup - The protocol setup configuration

msg_receiver - A closure that processes each message receiver