dkls23
Module proto

Misc protocol helper functions.

Re-exports

pub use tags::FilteredMsgRelay;

pub use tags::Round;

Modules

tags

tags

Structs

EncryptedMessage

A wrapper for a message of type T with support for in-place encryption/decryption with additional data.

MessageTag

MsgHdr

SignedMessage

A wrapper for a message of type T with support inplace signing and verifying

Traits

EncryptionScheme

Represents an encryption scheme interface for in-place AEAD and managing key exchange.

FixedExternalSize

A type with fixed size of external representation.

Relay

Wrap

A type with some external represention.

Functions

check_abort

Returns passed error if msg is a vaild abort message.

create_abort_message

Create an Abort Message.

decode_affine

Encode AffinePoint

decode_nonzero

Decode a NonZeroScalar

decode_point

Decode ProjectivePoint

decode_scalar

Decode a scalar

encode_affine

Encode AffinePoint

encode_point

Encode ProjectivePoint

encode_scalar

Encode a scalar

Type Aliases

PointBytes

External representation of a point on a curve

ScalarBytes

External Scalar representation

Scheme

Default encryption scheme