dkls23::proto

Struct EncryptedMessage

pub struct EncryptedMessage<T> { /* private fields */ }

A wrapper for a message of type T with support for in-place encryption/decryption with additional data.

Format of encrypted message:

[ msg-hdr | additional-data | payload | trailer | tag + nonce ]

payload | trailer are encrypted.

trailer is a variable-sized part of the message.

payload is the external representation of T.

Implementations

impl<T: AnyBitPattern + NoUninit> EncryptedMessage<T>

pub fn size(ad: usize, trailer: usize, scheme: &dyn EncryptionScheme) -> usize

Size of the whole message with additional data and trailer bytes.

pub fn new( id: &MsgId, ttl: u32, flags: u16, trailer: usize, scheme: &dyn EncryptionScheme, ) -> Self

Allocate a message with passed ID and TTL and additional trailer bytes.

pub fn new_with_ad( id: &MsgId, ttl: u32, flags: u16, additional_data: usize, trailer: usize, scheme: &dyn EncryptionScheme, ) -> Self

Allocate a message with passed ID and TTL and additional data and trailer bytes.

pub fn from_buffer( buffer: Vec<u8>, id: &MsgId, ttl: u32, flags: u16, additional_data: usize, trailer: usize, scheme: &dyn EncryptionScheme, ) -> Self

Use existing buffer but make sure it has the right size.

pub fn payload_with_ad( &mut self, scheme: &dyn EncryptionScheme, ) -> (&mut T, &mut [u8], &mut [u8])

Return a mutable references to message payload object, trailer and additional data byte slices.

pub fn payload(&mut self, scheme: &dyn EncryptionScheme) -> (&mut T, &mut [u8])

Return a mutable reference to message payload object and trailer byte slice.

pub fn encrypt( self, scheme: &mut dyn EncryptionScheme, receiver: usize, ) -> Option<Vec<u8>>

Encrypt message.

pub fn decrypt_with_ad<'msg>( buffer: &'msg mut [u8], additional_data: usize, trailer: usize, scheme: &dyn EncryptionScheme, sender: usize, ) -> Option<(&'msg T, &'msg [u8], &'msg [u8])>

Decrypt message and return references to the payload, trailer and additional data bytes.

pub fn decrypt<'msg>( buffer: &'msg mut [u8], trailer: usize, scheme: &dyn EncryptionScheme, sender: usize, ) -> Option<(&'msg T, &'msg [u8])>

Decrypte message and return reference to the payload and trailer bytes.