dkls23::proto::tags

Struct Round

pub struct Round<'a, R> { /* private fields */ }

Structure to receive a round of messages

Implementations

impl<'a, R: Relay> Round<'a, R>

pub fn new( count: usize, tag: MessageTag, relay: &'a mut FilteredMsgRelay<R>, ) -> Self

Create a new round with a given number of messages to receive.

pub async fn recv(&mut self) -> Result<Option<(Vec<u8>, usize, bool)>, Error>

Receive next message in the round. On success returns Ok(Some(message, party_index, is_abort_flag)). At the end of the round it returns Ok(None).

pub fn put_back(&mut self, msg: &[u8], tag: MessageTag, party_id: usize)

It is possible to receive a invalid message with a correct ID. In this case, it have to put the message id back into relay.expected table and increment a counter of waiting messages in the round.

pub async fn of_signed_messages<T, F, S, E>( self, setup: &S, abort_err: impl Fn(usize) -> E, handler: F, ) -> Result<(), E>
where T: AnyBitPattern + NoUninit, S: ProtocolParticipant, F: FnMut(&T, usize) -> Result<(), E>, E: From<Error>,

Receiver all messages in the round, verify, decode and pass to given handler.

pub async fn of_encrypted_messages<T, F, P, E>( self, setup: &P, scheme: &mut dyn EncryptionScheme, trailer: usize, err: impl Fn(usize) -> E, handler: F, ) -> Result<(), E>
where T: AnyBitPattern + NoUninit, P: ProtocolParticipant, F: FnMut(&T, usize, &[u8], &mut dyn EncryptionScheme) -> Result<Option<Vec<u8>>, E>, E: From<Error>,

Receiver all messages in the round, decrypt, decode and pass to given handler.

pub async fn broadcast_4<P, T1, T2, T3, T4>( self, setup: &P, msg: (T1, T2, T3, T4), ) -> Result<(Pairs<T1, usize>, Pairs<T2, usize>, Pairs<T3, usize>, Pairs<T4, usize>), Error>
where P: ProtocolParticipant, T1: Wrap, T2: Wrap, T3: Wrap, T4: Wrap,

Broadcast 4 values and collect 4 values from others in the round

pub async fn recv_broadcast_4<P, T1, T2, T3, T4>( self, setup: &P, sizes: &[usize; 4], ) -> Result<(Pairs<T1, usize>, Pairs<T2, usize>, Pairs<T3, usize>, Pairs<T4, usize>), Error>
where P: ProtocolParticipant, T1: Wrap, T2: Wrap, T3: Wrap, T4: Wrap,

Receive broadcasted 4 values from all parties in the round and collect them into 4 Pairs vectors.