dkls23::sign

Function message_receivers

pub fn message_receivers<S, F>(setup: &S, variant: DsgVariant, msg_receiver: F)
where
    S: ProtocolParticipant,
    F: FnMut(MsgId, &S::MessageVerifier),

Generates a map of message receivers for the DSG protocol

This function helps set up the message routing for the DSG protocol by determining which messages should be sent to which participants.

Type Parameters

S - A type implementing the ProtocolParticipant trait

F - A closure type for handling message receivers

Arguments

setup - The protocol setup configuration

variant - The variant of the DSG protocol to run

msg_receiver - A closure that will be called for each (message_id, verifier) pair