dkls23::proto::tags

Struct FilteredMsgRelay

pub struct FilteredMsgRelay<R> { /* private fields */ }

custom message relay

Implementations

impl<R: Relay> FilteredMsgRelay<R>

pub fn new(relay: R) -> Self

Construct a FilteredMsgRelay by wrapping up a Relay object

pub fn into_inner(self) -> R

Return underlying relay object

pub async fn expect_message( &mut self, id: MsgId, tag: MessageTag, party_id: usize, ttl: u32, ) -> Result<(), MessageSendError>

Mark message with ID as expected and associate pair (party-id, tag) with it.

pub async fn recv( &mut self, tag: MessageTag, ) -> Result<(Vec<u8>, usize, bool), Error>

Receive an expected message with given tag, and return a party-id associated with it.

pub async fn ask_messages<P: ProtocolParticipant>( &mut self, setup: &P, tag: MessageTag, p2p: bool, ) -> Result<usize, MessageSendError>

Add expected messages and Ask underlying message relay to receive them.

pub async fn ask_messages_from_iter<P, I>( &mut self, setup: &P, tag: MessageTag, from_parties: I, p2p: bool, ) -> Result<usize, MessageSendError>
where P: ProtocolParticipant, I: IntoIterator<Item = usize>,

Ask set of messages with a given tag from a set of parties.

Filter out own party_index from parties.

Returns number of messages with the same tag.

pub async fn ask_messages_from_slice<'a, P, I>( &mut self, setup: &P, tag: MessageTag, from_parties: I, p2p: bool, ) -> Result<usize, MessageSendError>
where P: ProtocolParticipant, I: IntoIterator<Item = &'a usize>,

The same as ask_messages_from_iter() by accepts slice of indices

pub fn round(&mut self, count: usize, tag: MessageTag) -> Round<'_, R>

Create a round