dkls23::proto
Trait Wrap

pub trait Wrap: Sized {
    // Required methods
    fn external_size(&self) -> usize;
    fn write(&self, buffer: &mut [u8]);
    fn read(buffer: &[u8]) -> Option<Self>;

    // Provided methods
    fn encode<'a>(&self, buf: &'a mut [u8]) -> &'a mut [u8]  { ... }
    fn decode(input: &[u8], size: usize) -> Option<(&[u8], Self)> { ... }
}

A type with some external represention.

Required Methods

fn external_size(&self) -> usize

Size of external representation in bytes

fn write(&self, buffer: &mut [u8])

Serialize a value into passed buffer

fn read(buffer: &[u8]) -> Option<Self>

Deserialize value from given buffer

Provided Methods

fn encode<'a>(&self, buf: &'a mut [u8]) -> &'a mut [u8]

Encode a value into passed buffer and return remaining bytes.

fn decode(input: &[u8], size: usize) -> Option<(&[u8], Self)>

Decode a value from input buffer using size bytes. Return remaining bytes and decoded value.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.