dkls23
Module setup

Setup message creation/parsing

Protocol setup message

Modules

finish

Setup for Finish PreSignature

key_export

Setup for Key export

keygen

Setup for DKG

quorum_change

Setup for Quorum Change

sign

Setup for DSG

Structs

AllOtherParties

An iterator for parties in range 0..total except me.

ArcSigner

Allow to pass Arc where expected SK: Signer

NoSignature

Type of empty signature.

NoSigningKey

NoVerifyingKey

A verifying key for NoSignature. Verification always succeeds. In this case verifying key used as an idenitity ID and communication uses a secure transport and there is no need to verify authenticity of the messages.

Constants

ABORT_MESSAGE_TAG

Tag of a broadcast message indicating that sender won’t participate in the protocol. The payload of the message contains error code.

SETUP_MESSAGE_TAG

Tag for all setup messages

Traits

FinalSignSetupMessage

A setup message for sign::finish()

KeyExportReceiverSetupMessage

A setup message for a reciever of exported key.

KeyExporterSetupMessage

A setup message for key export.

KeygenSetupMessage

A setup message for keygen::run()

PreSignSetupMessage

A setup message for sign::pre_signature()

ProtocolParticipant

Type that provides a protocol participant details.

QuorumChangeSetupMessage

A setup message for quorum_change::run()

SignSetupMessage

A setup message for sign::run()

SignatureEncoding

Support for decoding/encoding signatures as bytes.

Signer

Sign the provided message bytestring using Self (e.g. a cryptographic key or connection to an HSM), returning a digital signature.

Verifier

Verify the provided message bytestring using Self (e.g. a public key)